datn-xray-model-api/requirements.txt
flask
flask-restful
imageio
numpy==1.22.2
opencv-python
waitress
Pillow
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.3+cpu
